OAuth2Client: refactor ResourceOwnerEmulator
This PR refactors `ResourceOwnerEmulator` into
separate classes per grant type, which not
only reduces the class sizes, but also allows
to use more than one instance simultaneously,
in the same test.
